backend-python/requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.20.0
pillow==10.4.0
numpy==2.1.3

